ProgramNode#pathToRoot is now computed, not stored
Used to be that we passed each ProgramNode, at creation, a queue of
integers indicating the route you would have to take to get down to the
node in question.

However, this meant that we were storing this queue for every single
node in the population, including ones for which it would never matter,
which was the vast majority.

Hence we now compute that path when we need it, rather than store it off
at the time of instantiation, and as such we can complete the
11-multiplexer example without filling up the heap.